+++ crates/chr-reatlas/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
+## [Unreleased]
+
+### Removed
+
+- CLI (moved to reatlaser-cli)
+
 ## [0.1.1] - 2025-12-15
 
 ### Fixed
@@ -14,5 +20,6 @@ Based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 - Initial project
 
+[Unreleased]: https://github.com/yynaa/chr-reatlas
 [0.1.1]: https://github.com/yynaa/chr-reatlas/releases/tag/v0.1.1
 [0.1.0]: https://github.com/yynaa/chr-reatlas/releases/tag/v0.1.0